added auth check to firebase-actions

+++ src/lib/firebase-actions/deleteSharedBook.ts
@@ -1,12 +1,24 @@
 "use server";
 
 import { FieldValue } from "firebase-admin/firestore";
-import { adminDB } from "../firebase-admin";
+import { adminAuth, adminDB } from "../firebase-admin";
 import { revalidatePath } from "next/cache";
+import { cookies } from "next/headers";
 
 export async function deleteSharedBook(bookId: string, userId: string) {
   if (!adminDB) {
     throw new Error("Cannot connect to database");
+  }
+
+  if (!adminAuth) {
+    throw new Error("Cannot connect to auth database");
+  }
+
+  const idToken = cookies().get("idToken")?.value;
+  const decodedToken = idToken && (await adminAuth.verifyIdToken(idToken));
+
+  if (!decodedToken) {
+    throw new Error("Unauthorized");
   }
 
   await adminDB.collection("shared-books").doc(bookId).delete();
